requirements.txt
django==4.0.4
django-ckeditor==6.2.0